+++ src/MLF/Frontend/ConstraintGen.hs
@@ -8,7 +8,7 @@ module MLF.Frontend.ConstraintGen (
 import Data.Functor.Foldable (cata)
 import qualified Data.IntSet as IntSet
 
-import MLF.Frontend.Syntax (SurfaceExpr)
+import MLF.Frontend.Syntax (NormSurfaceExpr)
 import MLF.Frontend.Desugar (desugarSurface)
 import MLF.Constraint.Types.Graph (NodeId, PolySyms, cAnnEdges, getEdgeId)
 import MLF.Frontend.ConstraintGen.Types
@@ -92,7 +92,7 @@ Paper references:
     discusses the design choice of annotation-free let-polymorphism
 -}
 
-generateConstraints :: PolySyms -> SurfaceExpr -> Either ConstraintError ConstraintResult
+generateConstraints :: PolySyms -> NormSurfaceExpr -> Either ConstraintError ConstraintResult
 generateConstraints polySyms expr = do
     let expr' = desugarSurface expr
     let initialState = mkInitialStateWithPolySyms polySyms