+++ test/Main.hs
@@ -22,6 +22,7 @@ import qualified GraphOpsSpec
 import qualified CanonicalizerSpec
 import qualified FrontendParseSpec
 import qualified FrontendPrettySpec
+import qualified FrontendNormalizeSpec
 import qualified XMLFParseSpec
 import qualified XMLFPrettySpec
 
@@ -47,5 +48,6 @@ main = hspec $ do
     CanonicalizerSpec.spec
     FrontendParseSpec.spec
     FrontendPrettySpec.spec
+    FrontendNormalizeSpec.spec
     XMLFParseSpec.spec
     XMLFPrettySpec.spec